% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_art_map.r
\name{create_art_map}
\alias{create_art_map}
\title{Using Open Streetmaps API obtain information about Salamanca and plot a map}
\usage{
create_art_map()
}
\description{
Using Open Streetmaps API obtain information about Salamanca and plot a map
}
\examples{
create_art_map()
}
